split out projection, slice, and onedspectrum into their own module. We'll probably want to import these from somewhere else in the long term anyway
